+++ crates/miden-tx/src/host/account_procedures.rs
@@ -1,13 +1,13 @@
 use alloc::string::ToString;
 
+use miden_lib::transaction::TransactionAdviceInputs;
 use miden_lib::transaction::memory::{ACCOUNT_STACK_TOP_PTR, ACCT_CODE_COMMITMENT_OFFSET};
-use miden_lib::transaction::{TransactionAdviceInputs, TransactionKernelError};
 use miden_objects::account::{AccountCode, AccountProcedureInfo};
 use miden_objects::transaction::{TransactionArgs, TransactionInputs};
 use miden_processor::AdviceInputs;
 
 use super::{BTreeMap, Felt, ProcessState, Word};
-use crate::errors::TransactionHostError;
+use crate::errors::{TransactionHostError, TransactionKernelError};
 
 // ACCOUNT PROCEDURE INDEX MAP
 // ================================================================================================